Fix missing hour 0 by using Helsinki timezone boundaries for ENTSO-E requests
The issue was that ENTSO-E requests were using UTC boundaries (00:00-24:00 UTC)
instead of Helsinki local time boundaries. This caused the first hour of the day
(00:00-01:00 Helsinki time) to be missed because it starts at 22:00/23:00 UTC
the previous day.

Now properly converts Helsinki local time boundaries to UTC for the API request,
ensuring all 24 hours of the target date are included in the response.